-Added time library. Random seed now uses system clock. Needs test on non PC platforms.

+++ engine/source/randlib/rand32.h
@@ -9,6 +9,10 @@
 #ifndef RAND32_H
 #define RAND32_H
 
+#include<stdio.h>
+#include<stdlib.h>
+#include<time.h>
+
 // 2014-08-24, DC: Package seed into a stuct so we don't have
 // more unstructured global variables floating around.
 typedef struct
@@ -16,11 +20,9 @@ typedef struct
     unsigned long seed;
 } s_rand;
 
-//extern unsigned long seed;
-
 unsigned int rand32(void);
 
-void srand32(int n);
+void srand32(unsigned long n);
 
 #endif
